+++ src/pkg/shelf.rs
@@ -174,7 +174,6 @@ impl Shelf {
     }
 }
 
-
 #[cfg(test)]
 mod tests {
     use super::*;